+++ changelog.md
@@ -1,3 +1,15 @@
 # Changelog
 
-Opened 'ver3' branch.
+**./package.json**
+* Replaced 'ora' with 'ora-classic'
+
+---
+
+**Affected Files**
+* ./src/file-check.js
+* ./src/file-read.js
+* ./src/output-csv-files.js
+* ./src/output-folder.js
+* ./src/output-json-files.js
+* ./src/output-sql-files.js
+* ./src/proc-data.js